src/data/fetchers.py
"""
OpenEO data fetchers for Sentinel-2 and ERA5 weather data via Copernicus.

This module handles all data download operations from OpenEO backends.
Supports multiple backends:
- EODC (https://openeo.creo.vito.be) — European Open Data Cube
- CDSE (https://openeo.dataspace.copernicus.eu) — Copernicus Data Space Ecosystem (free)
- Sentinel Hub
"""

from datetime import datetime
from pathlib import Path
from typing import Optional, Tuple

import openeo
import xarray as xr
from loguru import logger


def authenticate_openeo(
    backend_url: str,
    username: str,
    password: str,
) -> openeo.Connection:
    """
    Authenticate with OpenEO backend.

    Args:
        backend_url: OpenEO backend URL (e.g., https://openeo.creo.vito.be/openeo/1.1.0)
        username: OpenEO username
        password: OpenEO password

    Returns:
        Authenticated openeo.Connection object

    Raises:
        Exception: If authentication fails

    Example:
        >>> conn = authenticate_openeo(
        ...     backend_url="https://openeo.creo.vito.be/openeo/1.1.0",
        ...     username="user@example.com",
        ...     password="password"
        ... )
    """
    try:
        connection = openeo.connect(backend_url)
        connection.authenticate_basic(username, password)
        logger.info(f"Authenticated with OpenEO backend: {backend_url}")
        return connection
    except Exception as e:
        logger.error(f"Failed to authenticate with OpenEO: {e}")
        raise


def fetch_sentinel2(
    backend_url: str,
    username: str,
    password: str,
    roi: list,
    start_date: str,
    end_date: str,
    max_cloud_cover: int = 20,
    bands: Optional[list] = None,
) -> openeo.DataCube:
    """
    Fetch Sentinel-2 imagery from OpenEO backend.

    Args:
        backend_url: OpenEO backend URL
        username: OpenEO username
        password: OpenEO password
        roi: Region of interest as [west, south, east, north]
        start_date: Start date in 'YYYY-MM-DD' format
        end_date: End date in 'YYYY-MM-DD' format
        max_cloud_cover: Maximum allowed cloud cover percentage (default 20)
        bands: List of band names to select (default: all S2 bands)

    Returns:
        openeo.DataCube with Sentinel-2 scenes

    Example:
        >>> roi = [2.754, 50.674, 5.918, 51.507]  # Flanders bbox
        >>> s2 = fetch_sentinel2(
        ...     backend_url, username, password,
        ...     roi, '2023-07-01', '2023-07-31'
        ... )
    """
    # TODO: Implement OpenEO Sentinel-2 collection fetch
    # - Authenticate with backend
    # - Create bbox from roi
    # - Load Sentinel-2 L2A collection
    # - Filter by date range and cloud cover (using SCL band)
    # - Select bands
    # - Return DataCube
    pass


def fetch_era5_daily(
    backend_url: str,
    username: str,
    password: str,
    roi: list,
    start_date: str,
    end_date: str,
    variables: Optional[list] = None,
) -> openeo.DataCube:
    """
    Fetch ERA5 daily weather data from OpenEO backend.

    Args:
        backend_url: OpenEO backend URL
        username: OpenEO username
        password: OpenEO password
        roi: Region of interest as [west, south, east, north]
        start_date: Start date in 'YYYY-MM-DD' format
        end_date: End date in 'YYYY-MM-DD' format
        variables: List of variable names to fetch
            (e.g., 'temperature_2m', 'total_precipitation_sum')

    Returns:
        openeo.DataCube with ERA5 daily climate variables

    Example:
        >>> era5 = fetch_era5_daily(
        ...     backend_url, username, password,
        ...     roi, '2023-07-01', '2023-07-31'
        ... )
    """
    # TODO: Implement ERA5 data fetch via OpenEO
    # - Authenticate
    # - Load ERA5 collection
    # - Filter by date range, spatial extent
    # - Select variables
    # - Return DataCube
    pass


def download_to_netcdf(
    datacube: openeo.DataCube,
    output_file: str,
    format: str = "netcdf",
) -> str:
    """
    Download OpenEO DataCube to NetCDF file.

    Args:
        datacube: openeo.DataCube to download
        output_file: Output file path
        format: Output format ('netcdf', 'GeoTIFF', 'COG', etc.)

    Returns:
        Path to downloaded file

    Note:
        Uses synchronous download (download_result())
    """
    # TODO: Implement OpenEO download to NetCDF
    # - Prepare output format options
    # - Execute synchronous download
    # - Validate output
    # - Return file path
    pass


def download_to_geotiff(
    datacube: openeo.DataCube,
    output_file: str,
) -> str:
    """
    Download OpenEO DataCube to Cloud Optimized GeoTIFF (COG).

    Args:
        datacube: openeo.DataCube to download
        output_file: Output file path

    Returns:
        Path to downloaded file
    """
    # TODO: Implement COG download
    pass
