pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vegetation-drought-resilience"
version = "0.1.0"
description = "Drought impact monitoring using satellite imagery and weather data"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Vegetation Drought Resilience Team", email = "team@example.com"},
]
keywords = ["drought", "monitoring", "satellite", "imagery", "weather", "geospatial"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "xarray>=2023.6.0",
    "rasterio>=1.3.0",
    "geopandas>=0.13.0",
    "rioxarray>=0.14.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    "geoalchemy2>=0.13.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "folium>=0.14.0",
    "plotly>=5.15.0",
    "streamlit>=1.27.0",
    "streamlit-folium>=0.7.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "pytest-xdist>=3.3.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

docs = [
    "sphinx>=6.2.0",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-autodoc-typehints>=1.23.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/vegetation_drought_resilience"
Documentation = "https://github.com/yourusername/vegetation_drought_resilience/tree/main/docs"
Repository = "https://github.com/yourusername/vegetation_drought_resilience.git"
"Bug Tracker" = "https://github.com/yourusername/vegetation_drought_resilience/issues"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_first_party = ["src", "app", "tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing --verbose"
markers = [
    "integration: marks tests as integration tests",
    "slow: marks tests as slow",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
